% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gampt.R
\name{gampt.infil}
\alias{gampt.infil}
\title{Calculates Green-Ampt infiltration for a given rainfall timeseries
Reference: WinGampt by Parson & Carpena (2009): Based largely on this fortran program.
http://abe.ufl.edu/carpena/software/wingampt.shtml}
\usage{
gampt.infil(Time, rainfall, soil.para, smax = 0.5, tstep, tol = 1e-06, ...)
}
\arguments{
\item{Time}{Time input}

\item{rainfall}{Rainfall vector corresponding to Time input}

\item{soil.para}{Soil hydraulic parameters:
saturated water content(ths), initial water content(thi), Ksat, Sav}

\item{smax}{Maximum storage at the surface}

\item{tol}{Maximum tolerance for the absolute error in iterative solution of Chu's equation}
}
\description{
Calculates Green-Ampt infiltration for a given rainfall timeseries
Reference: WinGampt by Parson & Carpena (2009): Based largely on this fortran program.
http://abe.ufl.edu/carpena/software/wingampt.shtml
}
\examples{

infil = gampt.infil(Time = seq(0, 4, 0.1),
rainfall = c(0, rep(1.5, 10), rep(0.1, 10), rep(1.0, 20)),
soil.para = list(ths = 0.499, thi = 0.25, Ksat = 0.044, Sav = 22.4),
smax = 0.8, tstep = 0.1, tol = 0.000001)

infil$water_balance
}
