% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtnp.R
\name{newtnp}
\alias{newtnp}
\title{Iterative solution to calculate end of ponding time}
\usage{
newtnp(tstart, tend, tp, tpp, rfi, amtinf, soil.para, tol = 1e-05, ...)
}
\arguments{
\item{tstart}{Start of event}

\item{tend}{End of event}

\item{tp}{Ponding time}

\item{tpp}{Equaivalent time for initially ponded condition}

\item{soil.para}{Soil hydraulic parameters}

\item{tol}{Tolerance threshold for abslolute error in iterative solution}
}
\description{
Iterative solution to calculate end of ponding time
}
