% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sschu.R
\name{sschu}
\alias{sschu}
\title{Chu (1988)' iterative soilution}
\usage{
sschu(Fpi, Time, tp, tpp, soil.para, tol = 1e-05, ...)
}
\arguments{
\item{Fpi}{Initial value for cumulative infiltration}

\item{Time}{Time}

\item{tp}{Ponding time}

\item{tpp}{Equaivalent time for initially ponded condition}

\item{soil.para}{Soil hydraulic parameter}

\item{tol}{Tolerance threshold for abslolute error in iterative solution}
}
\description{
Chu (1988)' iterative soilution
}
